Some CORBA cleanups, fix problems with unsigned types, start exploring how to support derived types.
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@681597 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/bindings/corba/src/main/java/org/apache/cxf/binding/corba/types/CorbaPrimitiveHandler.java
@@ -191,7 +191,7 @@ public class CorbaPrimitiveHandler extends CorbaObjectHandler {
             obj = new Character(data.charAt(0));
             break;
         case TCKind._tk_octet:
-            obj = new Byte(data);
+            obj = new Short(data).byteValue();
             break;
         case TCKind._tk_short:
             obj = new Short(data);
